qwen3_tts_nv: update scripts and dir structure after rebasing to 0.21
